Refactored the project structure
CSVReaderUtil: load the CSV file and map it to UserBean

UserBean: Bean class for CSV mapping and will be used by DataProviders

Moved the test data from src/main/resources to src/test/resources.
